requirements.txt
streamlit
Pillow
pytesseract


